+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 
 ### Bug fixes
 
+- Fixed an issue where renaming a profile with token authentication caused an infinite 401 credential prompt loop. The authentication state is now properly transferred when a profile is renamed. [#3784](https://github.com/zowe/zowe-explorer-vscode/issues/3784)
 - Fixed an issue where secure credentials and headers were being logged to the Zowe logger and VSCode output channel. [#3848](https://github.com/zowe/zowe-explorer-vscode/pull/3848)
 - Updated Zowe SDKs to version `8.28.0` to address an issue where copying a PDS member to a data set across LPARs failed. This occurred when the target PDS already contained members, but none matched the name of the PDS member being copied. [#3848](https://github.com/zowe/zowe-explorer-vscode/pull/3896)
 - Ensure that the `updateCredentials` function adds the given credentials to the `availableCreds` cache of the `Session` object. [#3940](https://github.com/zowe/zowe-explorer-vscode/pull/3940)